+++ CODE_OF_CONDUCT.md
@@ -1,12 +1,12 @@
 # 📜 Code of Conduct
 
-## Contributor Covenant — Adapted for Robo-Tech Forum
+## Contributor Covenant — Adapted for THE  ROBO-TECH FORUM
 
 ---
 
 ## Our Pledge
 
-We as members, contributors, and maintainers of the Robo-Tech Forum (RTF) website project pledge to make participation in our community a **harassment-free experience for everyone**, regardless of age, body size, visible or invisible disability, ethnicity, sex characteristics, gender identity and expression, level of experience, education, socio-economic status, nationality, personal appearance, race, caste, religion, or sexual identity and orientation.
+We as members, contributors, and maintainers of the THE  ROBO-TECH FORUM (RTF) website project pledge to make participation in our community a **harassment-free experience for everyone**, regardless of age, body size, visible or invisible disability, ethnicity, sex characteristics, gender identity and expression, level of experience, education, socio-economic status, nationality, personal appearance, race, caste, religion, or sexual identity and orientation.
 
 We pledge to act and interact in ways that contribute to an **open, welcoming, diverse, inclusive, and healthy community**.
