+++ index.html
@@ -4,11 +4,11 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/png" href="/rtf-logo-img.jpg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="Robo-Tech Forum (RTF) — The premier robotics club at Government College of Engineering, Amravati. Engineering excellence, built by students, recognized by industry." />
+    <meta name="description" content="THE  ROBO-TECH FORUM (RTF) — The premier robotics club at Government College of Engineering, Amravati. Engineering excellence, built by students, recognized by industry." />
     <meta name="theme-color" content="#020408" />
 
     <!-- Open Graph -->
-    <meta property="og:title" content="Robo-Tech Forum — GCoEA Amravati" />
+    <meta property="og:title" content="THE  ROBO-TECH FORUM — GCoEA Amravati" />
     <meta property="og:description" content="Engineering excellence. Built by students. Recognized by industry." />
     <meta property="og:image" content="/og-image.png" />
     <meta property="og:type" content="website" />
@@ -21,7 +21,7 @@
       rel="stylesheet"
     />
 
-    <title>Robo-Tech Forum — GCoEA Amravati</title>
+    <title>THE  ROBO-TECH FORUM — GCoEA Amravati</title>
   </head>
   <body>
     <!-- Skip to content — Accessibility -->
