+++ src/App.jsx
@@ -1,14 +1,12 @@
 import { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom';
 import { AnimatePresence } from 'framer-motion';
-import { useLenis } from './hooks/useLenis';
 import { useLoadingManager } from './hooks/useLoadingManager';
 import { LoadingProvider, useLoading } from './context/LoadingContext';
 
 // Layout
 import Navbar from './components/layout/Navbar';
 import Footer from './components/layout/Footer';
-import ScrollProgressFixed from './components/layout/ScrollProgressFixed';
 import VideoIntroEnhanced from './components/layout/VideoIntroEnhanced';
 import LoadingScreen from './components/layout/LoadingScreen';
 
@@ -51,9 +49,6 @@ function AppContent() {
   const { shouldShowIntro, markIntroComplete, assetsReady, setAssetsReady } = useLoadingManager();
   const { isLoading, loadingMessage } = useLoading();
 
-  // Buttery smooth scrolling via Lenis, synced to GSAP ScrollTrigger
-  useLenis();
-
   // Handle scroll to top button
   useEffect(() => {
     const handleScroll = () => {
@@ -88,7 +83,6 @@ function AppContent() {
 
       {/* Main Content */}
       <div className="relative min-h-screen bg-deep text-text-primary" style={{ overflowX: 'clip' }}>
-        <ScrollProgressFixed />
         <Navbar />
         <AnimatedRoutes />
         <Footer />
