+++ src/pages/Home.jsx
@@ -6,17 +6,6 @@ import FeaturedProjects from '../components/sections/FeaturedProjects';
 import ParallaxImage from '../components/ui/ParallaxImage';
 import TerminalContact from '../components/sections/TerminalContact';
 
-/**
- * Home page
- *
- * Layout trick that replicates the IITD reference site effect:
- *   1. <HeroSection> renders a 280vh tall div with the hero PINNED inside it.
- *      The hero sits at z-index:1 and stays stuck while browser scrolls.
- *
- *   2. .overlay-panel below has z-index:20 + dark bg + rounded-top corners.
- *      As the user scrolls past the hero's 280vh runway, this panel slides
- *      UP and OVER the pinned hero — creating the "covering" cinematic effect.
- */
 export default function Home() {
   return (
     <motion.main
@@ -26,19 +15,9 @@ export default function Home() {
       animate="animate"
       exit="exit"
     >
-      {/* ── PINNED HERO — 280vh scroll runway ── */}
       <HeroSection />
 
-      {/* ── OVERLAY PANEL — slides over the pinned hero ── */}
-      <div
-        className="relative bg-deep overlay-panel"
-        style={{
-          zIndex: 20,
-          borderRadius: '28px 28px 0 0',
-          marginTop: '-1px',          // flush seam
-          boxShadow: '0 -30px 80px rgba(0,0,0,0.85)',
-        }}
-      >
+      <div className="relative bg-deep z-20">
         <StatsBar />
 
         {/* Parallax break — workshop / lab image */}
@@ -47,7 +26,6 @@ export default function Home() {
           images={[
             'https://therobotechforum.in/assets/img/Robocon%202025/20250713_183423%20-%20Copy.webp',
             'https://therobotechforum.in/assets/img/College%20Programs/20241012_085453.webp',
-            // 'https://therobotechforum.in/assets/img/Robocon%202025/IMG_20250713_183012.webp',
           ]}
           alt="Robotics workshop"
           overlay="We don't just study engineering — we live it."
